Refactor ingest and graph helper surface
Split FILE-block parsing, PDF ingest helpers, and graph visual helpers into focused owner modules while preserving existing public exports through ingest.ts and graph visual contract coverage.

Keep live Gemini, real Vault ingest, and cleanup/delete out of this change; only fixture dry-run smoke was run.

Tested: git diff --check; npm run typecheck; npm run test:mocks -- src/components/graph/graph-view-visual-contract.test.ts src/lib/ingest-parse.test.ts src/lib/ingest.prompt.test.ts; npm test; npm run smoke:live-ingest -- --fixture --dry-run.
